Enable all compile warnings, and fix some of them.

+++ src/cfrds_buffer.c
@@ -472,7 +472,7 @@ bool cfrds_buffer_parse_bytearray(const char **data, size_t *remaining, char **o
 
 bool cfrds_buffer_parse_string(const char **data, size_t *remaining, char **out)
 {
-    int64_t size = 0;
+    size_t size = 0;
     int64_t tmp = 0;
 
     if (out == NULL)
